Made aa_tables into a class.
New class is AATables. It automatically reads in a fasta alignment and can
convert to biochem tables and write them out. It still needs cleaning up and
checking. The particular features to be checked are flexibility and efficiency,
since the functions write_tables and make_biochem kind of repeat some
functionality, and also how it deals with user defined tables.